requirements.txt
streamlit
pytube
openai-whisper
ffmpeg-python
opencv-python
keybert
torch
tqdm
transformers
scikit-learn
sentence-transformers
